_server/requirements.txt
pyserial==3.5
pywin32==310
